+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Update second_page to config_page.
+* Include default data if no Datacat object passed into config_page.
+
 ## 0.0.1
 
 * Initial config ui.